backend/requirements.txt
fastapi
uvicorn
boto3
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose
